tests/wpt/tests/navigation-api/navigation-history-entry/key-id-location-replace.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe id="i" src="/common/blank.html"></iframe>
<script>
async_test(t => {
  let start_length = navigation.entries().length;
  window.onload = t.step_func(() => {
    let key = i.contentWindow.navigation.currentEntry.key;
    let id = i.contentWindow.navigation.currentEntry.id;
    i.contentWindow.location.replace("/common/blank.html?query");
    i.onload = t.step_func_done(() => {
      assert_equals(key, i.contentWindow.navigation.currentEntry.key);
      assert_not_equals(id, i.contentWindow.navigation.currentEntry.id);
      assert_equals(navigation.entries().length, start_length);
    });
  });
}, "NavigationHistoryEntry's key and id after location.replace()");
</script>
